fix: 상품 등록/수정 데이터 수집 로직 커스텀 훅으로 분리
