+++ src/components/items/add/InputContainer.tsx
@@ -1,93 +1,23 @@
-import { ChangeEvent, KeyboardEvent, useEffect, useRef, useState } from "react";
 import DeleteBtnImg from "@/public/assets/images/items/cancel.svg";
 import { InitialValues } from "../shared/types";
 import styles from "../shared/form.module.css";
+import { useProductFormState } from "@/src/hooks/use/useProductFormState";
 
 // 상품 정보 등록
-function InputContainer({ setValues }: InitialValues) {
-  const [tag, setTag] = useState("");
-  const [tagList, setTagList] = useState<string[]>([]);
-  const [enteredNum, setEnterdNum] = useState<string | number>("");
-  const [price, setPrice] = useState<number>(0);
-  const priceInput = useRef<HTMLInputElement | null>(null);
-
-  useEffect(() => {
-    if (typeof enteredNum === "string") {
-      const numberPrice = enteredNum.replace(/,/g, "");
-      setPrice(Number(numberPrice));
-    }
-    if (priceInput.current)
-      if (priceInput.current.value === "") {
-        setPrice(0);
-      }
-
-    setValues((prevValue) => ({
-      ...prevValue,
-      ["price"]: price,
-    }));
-  }, [enteredNum, price, setValues]);
-
-  const handleChange = (name: string, value: string | number | string[]) => {
-    setValues((prevValue) => ({
-      ...prevValue,
-      [name]: value,
-    }));
-  };
-
-  const handleInputChange = (
-    e: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-  ) => {
-    let { name, value }: { name: string; value: string | number | string[] } =
-      e.target;
-    handleChange(name, value);
-  };
-
-  const changeEnteredNum = (e: ChangeEvent<HTMLInputElement>) => {
-    const value: string = e.target.value;
-    const removedCommaValue = Number(value.replaceAll(",", ""));
-    setEnterdNum(removedCommaValue.toLocaleString());
-
-    if (isNaN(removedCommaValue) || removedCommaValue === 0) {
-      setEnterdNum("");
-    }
-  };
-
-  // onKeyDown 이벤트 키가 Enter와 일치하면 실행
-  const activeEnter = (e: KeyboardEvent<HTMLInputElement>) => {
-    const specialCharRegex =
-      /[ \{\}\[\]\/?.,;:|\)*~`!^\+┼<>@\#$%&\'\"\\\(\=]/gi;
-    // onKeyDown 이벤트의 한글 입력 시 이벤트가 두 번 호출 되는 버그 방지
-    if (e.nativeEvent.isComposing) {
-      return;
-    }
-    // 특수문자 및 스페이스바 입력 방지
-    if (specialCharRegex.test(e.key)) {
-      e.preventDefault();
-    }
-    if (e.key === "Enter") {
-      if ((e.target as HTMLInputElement).value === "") {
-        // 빈 칸 엔터 방지
-        e.preventDefault();
-      } else {
-        handleChange("tag", tagList);
-        tagList.push(tag);
-        (e.target as HTMLInputElement).value = "";
-      }
-    }
-  };
-
-  const handleAddValue = (e: ChangeEvent<HTMLInputElement>) => {
-    setTag(e.target.value);
-  };
-
-  // 클릭한 태그 삭제
-  const handleDeleteClick = (i: number) => {
-    tagList.splice(i, 1);
-    handleChange("tag", tagList);
-  };
+function InputContainer({ setValues, uploadLoading }: InitialValues) {
+  const {
+    tagList,
+    enteredNum,
+    changeEnteredNum,
+    priceInput,
+    activeEnter,
+    handleInputChange,
+    handleAddValue,
+    handleDeleteClick,
+  } = useProductFormState(setValues);
 
   return (
-    <>
+    <fieldset disabled={uploadLoading}>
       <div className={styles.inputContainer}>
         <p className={styles.inputTitle}>상품명</p>
         <input
@@ -130,7 +60,6 @@ function InputContainer({ setValues }: InitialValues) {
         />
         <ul className={styles.tagContainer}>
           {tagList &&
-            // key 값으로 사용할 고유한 값이 없으므로 index 값 임시 부여
             tagList.map((item, i) => {
               return (
                 <li key={i}>
@@ -146,7 +75,7 @@ function InputContainer({ setValues }: InitialValues) {
             })}
         </ul>
       </div>
-    </>
+    </fieldset>
   );
 }
