+++ styles/app/reset.css
@@ -19,20 +19,17 @@
   font-weight: 300;
 }
 @font-face {
-  src: url("/assets/fonts/pretendard/Pretendard-Regular.woff")
-    format("woff");
+  src: url("/assets/fonts/pretendard/Pretendard-Regular.woff") format("woff");
   font-family: "pretendard";
   font-weight: 400;
 }
 @font-face {
-  src: url("/assets/fonts/pretendard/Pretendard-Medium.woff")
-    format("woff");
+  src: url("/assets/fonts/pretendard/Pretendard-Medium.woff") format("woff");
   font-family: "pretendard";
   font-weight: 500;
 }
 @font-face {
-  src: url("/assets/fonts/pretendard/Pretendard-SemiBold.woff")
-    format("woff");
+  src: url("/assets/fonts/pretendard/Pretendard-SemiBold.woff") format("woff");
   font-family: "pretendard";
   font-weight: 600;
 }
@@ -237,6 +234,10 @@ input:-webkit-autofill:hover,
 input:-webkit-autofill:focus,
 input:-webkit-autofill:active {
   transition: background-color 5000s ease-in-out 0s;
+}
+input:disabled,
+textarea:disabled {
+  cursor: not-allowed;
 }
 body {
   width: 100%;
